+++ src/App.js
@@ -4,10 +4,11 @@ import { ChakraProvider } from '@chakra-ui/react';
 import Header from './component/Header';
 import Write from './pages/Write';
 import Main from './pages/Main';
-import Join from './pages/Join';
 import My from './pages/My';
+import Sign from './pages/Sign';
 
 function App() {
+  const [loginButton, setLoginButton] = React.useState(false);
   return (
     <ChakraProvider>
       <Header />
@@ -15,7 +16,12 @@ function App() {
         <Routes>
           <Route path="/portfolio" element={<Write />} />
           <Route path="/" element={<Main isLogin={false} />} />
-          <Route path="/join" element={<Join isLogin={false} />} />
+          <Route
+            path="/sign"
+            element={
+              <Sign loginButton={loginButton} setLoginButton={setLoginButton} />
+            }
+          />
           <Route path="/mypage" element={<My />} />
         </Routes>
       </Router>
